+++ src/containerapp/azext_containerapp/_constants.py
@@ -143,3 +143,20 @@ SUPPORTED_RUNTIME_LIST = [RUNTIME_GENERIC, RUNTIME_JAVA]
 
 AKS_AZURE_LOCAL_DISTRO = "AksAzureLocal"
 SETUP_CORE_DNS_SUPPORTED_DISTRO = [AKS_AZURE_LOCAL_DISTRO]
+CUSTOM_CORE_DNS_VOLUME_NAME = 'custom-config-volume'
+CUSTOM_CORE_DNS_VOLUME_MOUNT_PATH = '/etc/coredns/custom'
+CUSTOM_CORE_DNS = 'coredns-custom'
+CORE_DNS = 'coredns'
+KUBE_SYSTEM = 'kube-system'
+EMPTY_CUSTOM_CORE_DNS = """
+apiVersion: v1
+data:
+kind: ConfigMap
+metadata:
+  labels:
+    addonmanager.kubernetes.io/mode: EnsureExists
+    k8s-app: kube-dns
+    kubernetes.io/cluster-service: "true"
+  name: coredns-custom
+  namespace: kube-system
+"""
