+++ src/containerapp/setup.py
@@ -47,7 +47,7 @@ CLASSIFIERS = [
 ]
 
 # TODO: Add any additional SDK dependencies here
-DEPENDENCIES = ['pycomposefile>=0.0.29', 'docker']
+DEPENDENCIES = ['pycomposefile>=0.0.29', 'docker', 'kubernetes==24.2.0']
 
 # Install pack CLI to build runnable application images from source
 _ = get_pack_exec_path()
